fix: use queueMicrotask in supported browsers
