+++ packages/@lwc/engine-core/src/framework/vm.ts
@@ -26,7 +26,7 @@ import { logError } from '../shared/logger';
 
 import { HostNode, HostElement, RendererAPI } from './renderer';
 import { renderComponent, markComponentAsDirty, getTemplateReactiveObserver } from './component';
-import { addCallbackToNextTick, EmptyArray, EmptyObject, flattenStylesheets } from './utils';
+import { EmptyArray, EmptyObject, flattenStylesheets } from './utils';
 import { invokeServiceHook, Services } from './services';
 import { invokeComponentCallback, invokeComponentConstructor } from './invoker';
 import { Template } from './template';
@@ -45,6 +45,7 @@ import { connectWireAdapters, disconnectWireAdapters, installWireAdapters } from
 import { removeActiveVM } from './hot-swaps';
 import { VNodes, VCustomElement, VNode, VNodeType, VBaseElement } from './vnodes';
 import { StylesheetFactory, TemplateStylesheetFactories } from './stylesheet';
+import { queueMicrotask } from './queue-microtask';
 
 type ShadowRootMode = 'open' | 'closed';
 
@@ -575,7 +576,7 @@ function flushRehydrationQueue() {
             if (i + 1 < len) {
                 // pieces of the queue are still pending to be rehydrated, those should have priority
                 if (rehydrateQueue.length === 0) {
-                    addCallbackToNextTick(flushRehydrationQueue);
+                    queueMicrotask(flushRehydrationQueue);
                 }
                 ArrayUnshift.apply(rehydrateQueue, ArraySlice.call(vms, i + 1));
             }
@@ -740,7 +741,7 @@ export function scheduleRehydration(vm: VM) {
 
     vm.isScheduled = true;
     if (rehydrateQueue.length === 0) {
-        addCallbackToNextTick(flushRehydrationQueue);
+        queueMicrotask(flushRehydrationQueue);
     }
 
     ArrayPush.call(rehydrateQueue, vm);
